more preparatory work to split NormalEquations into two classes. An abstract class derived from Equation (ImagingEquation) has been written to encapsulate functionality of type checking for the associated NormalEquation. I have a feeling that the responsibilities between various Equation classes are very vaguely defined. Some redesigning is probably necessary. In addition, there seems to be a problem with constness in scimath::Equation. However, for now some workarounds are written and the current state is left intact. The old NormalEquation is still used instead of the imaging-specific normal equations class, untill it is written. All tests/classes, which rely on the normal equations in the generic form have been converted to use GenericNormalEquations.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@865 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Synthesis/synthesis/trunk/parallel/ImagerParallel.cc
@@ -94,7 +94,7 @@ namespace conrad
 					}
 				}
 				if(itsNNode>1) {
-				  CONRADCHECK(itsMs.size()==(itsNNode-1),
+				  CONRADCHECK(int(itsMs.size()) == (itsNNode-1),
 					      "When running in parallel, need one data set per node");
 				}
 
@@ -152,7 +152,7 @@ namespace conrad
 					CONRADCHECK(itsSolver, "Solver not defined correctly");
 					itsSolver->init();
 					itsSolver->setParameters(*itsModel);
-					for (int iMs=0; iMs<itsMs.size(); iMs++)
+					for (size_t iMs=0; iMs<itsMs.size(); iMs++)
 					{
 						calcOne(itsMs[iMs]);
 						itsSolver->addNormalEquations(*itsNe);
